Add files via upload
The source files inside SRC_TFT are necessary to drive a touch screen TFT 3.5'' 480x320 with SPI, including the IC 9488.
The source files inside SRC_GUI are necessary to use the components (buttons, textboxes... ) for the graphical interface GUI.
The source files inside SRC_INC_myApp are related to my application (a squash scoreboard) and can be used as a template for your design.
The source files inside Graph_Win32_GUI are wrapper files related to my applications. This folder must be disregarded when using a microcontroller IDE. These files must be included in a Win32 project for your desktop IDE (e.g. VS2010). Please exclude prWin32.cpp, but use it as a template as the main file. If your project name is myProject, then  you should copy the lines of prWin32.cpp inside your myProject.cpp and substitute the following strings: prWin32 by myProject    and     PRWIN32 by MYPROJECT
The header files inside INC_TFT_GUI are all of the include files. Sorry for grouping all of them. This way is faster to include the folder in a IDE.